--- lib/Screens/widgets/cameraWidget.dart
@@ -1,32 +1,0 @@
-import 'package:camera/camera.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-class cameraWidget extends StatelessWidget {
-  const  cameraWidget({
-
-    required this.cameraValue,
-    required CameraController cameraController,
-  }) : _cameraController = cameraController;
-
-  final Future<void> cameraValue;
-  final CameraController _cameraController;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(top:25.0),
-      child: FutureBuilder(
-          future: cameraValue,
-          builder: (context, snapshot) {
-            if (snapshot.connectionState == ConnectionState.done) {
-              return CameraPreview(_cameraController);
-            } else {
-              return Center(
-                child: CircularProgressIndicator(),
-              );
-            }
-          }),
-    );
-  }
-}
