+++ lib/main.dart
@@ -2,13 +2,12 @@ import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:sample/Screens/currencyDenomination/currencyDenominationScreen.dart';
-import 'package:sample/Screens/currencyDenomination/showCurrency.dart';
-import 'package:sample/Screens/homeScreen/homeScreen.dart';
+import 'package:sample/Screens/currencyDenomination/currencyDenominations.dart';
+
 import 'package:sample/Screens/objectRecogination/objectRecogination.dart';
 import 'package:sample/Screens/onBoarding/onboardingScreen.dart';
 import 'package:sample/Screens/settings/settings.dart';
 import 'package:sample/Screens/splashScreen/splashScreen.dart';
-import 'package:sample/Screens/obstacleDetection/obstacleDetection.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -30,11 +29,10 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       home: SplashScreenPage(),
       routes: {
-        '/homeScreen': (context) => HomeScreen(),
         '/onboarding': (context) => OnboardingScreen(),
         //  '/obstacleDetection':
         '/currencyDenomination': (context) =>
-            currencyDetection(cameras: cameras),
+            CurrencyDenomination(cameras: cameras),
         '/objectRecogination': (context) => ObjectRecognition(cameras: cameras),
         '/settings': (context) => const Settings(),
       },
